reader/requirements.txt
kafka_python==2.0.2
python-dotenv==1.0.0
